more work on game_logic
Mainly added a ton of functions to game_logic.py for handling some of the major events of the game. I'm still figuring out the full structure of how the queue will be set up and handled. At this point, I really just need a stack for handling chains of events. A lot of things will be processed by the game on the inside by sending small events to handle to the stack over and over again in this format.
Sometimes, we will need to be able to tell if an event on the stack has a player's choice given already or if we need to create a node in the tree to give the player a choice. To handle this, I currently have a function that checks for both cases; creating a choice event if we don't have enough information or carrying out the event if we do.

Other changes:
- Added more lists of constants for functions to iterate over
- Changed the names of some functions