higher level functions
Putting in a few days-worth of work now, moving through more complex, upper-level functions now. The most tricky part now is managing costs. I have a new function that can easily tell if a player can afford a cost, but coming up with payments for the cost is trickier. I am in the middle of a cost-generating function, but just got the idea to pay for it item-by-item, instead of generating all possible payments in one go.

- Mostly finished the progress_game function, which automatically handles events until a choice or random event is added to the queue. This causes a new node to be added to our eventually-used game tree