Logic completed - 1st pass
I have finished pretty much all of the logic functions needed to run the entire game. NONE of it has actually been tested yet; this could also take a while, but I'm just glad to be done with this part.

I also want to work on the logging and/or visualization of games, but that has always been a tedious process. Perhaps I'll try using Copilot Agent or Edit to do this quickly.

Other changes:
- Main turn actions that require a coin or egg payment will have the cost added to the action itself in the beginning
- Trying to be proactive with making deep copies of dicts, so we don't edit actions or states that are in past nodes later on
- For now, the Automa just passes its turn
- Added a generator for random event outcomes
- Managing the "coords" key for actions and events has been the most confusing thing in the project so far, it's likely not perfect at the moment
- Added function to find out what major actions a player can do at the start of their turn
- Added functions that calculate the valid combinations possible of resources, egg locations, and cards to pay a cost. Payments of costs are handled one item at a time for simplicity.