fixes and end game abilities
Implemented the calculation of points from end game abilities. Added a cleanup pruning method to nodes to make memory much more manageable when running MCTS, with unvisited branches and paths being removed automatically by the garbage collector.
I also tried implementing a UCB algorithm from a paper in my bandit algorithms class, but it initially seems to only be effective with the current parameters with a very large number of samples.

Added the complete logs for a game played where a score of 84 points was achieved.

Other changes:
- Many abilities were updated to give the player a choice of whether to activate them or not instead of simply activating if the player could afford them.
- The setup choices (which 2 cards to discard and which 3 resources to take) are now condensed into one action for each (2 choices total), with each combination considered in one step. I may also do this with the refilling of the display cards in the future, but it doesn't seem necessary at the moment.
- Fixed bugs with the draw decision and resource decision events where the original event dictionaries were edited instead of new copies made for future states.