lots more test games
Ran many more test games over a few days time, comparing a few values of the UCT constant. It seems that some seem better than others, but it is not highly consistent. As a basic guess, the effectiveness of certain constants seems to relate to the number of sample games available, at least over the course of an entire game.

Implemented a version of the Nested Rollout Policy Algorithm (NRPA), though it does not work perfectly for games with chance nodes. It surely performs at a weaker level than UCT, but still does better than I expected it to.

Other changes:
- Improved logging in game_base_uct.py
- Started running some UCT games where the algorithm has a minimum of 2500 simulations to use per choice node